refactor(client)!: moving the `sn_interface::types::connections` module to the `sn_client` crate
- This module doesn't belong to the `sn_interface` crate since it's not
part of the network's protocol/API.
