+++ sn_client/src/connections/link.rs
@@ -8,19 +8,14 @@
 
 use super::Peer;
 
-use crate::messaging::MsgId;
-use crate::types::log_markers::LogMarker;
 use qp2p::{Connection, Endpoint, RecvStream, UsrMsgBytes};
-// Required for docs
-#[allow(unused_imports)]
-use qp2p::RetryConfig;
+use sn_interface::{messaging::MsgId, types::log_markers::LogMarker};
 use std::{collections::BTreeMap, sync::Arc};
 use tokio::sync::RwLock;
-use xor_name::XorName;
 
 type ConnId = String;
 
-/// A link to a peer in our network.
+/// A link to a peer in the network.
 ///
 /// The upper layers will add incoming connections to the link,
 /// and use the link to send msgs.
@@ -28,10 +23,8 @@ type ConnId = String;
 /// The link is a way to keep connections to a peer in one place
 /// and use them efficiently; converge to a single one regardless of concurrent
 /// comms initiation between the peers, and so on.
-/// Unused connections will expire, so the Link is cheap to keep around.
-/// The Link is kept around as long as the peer is deemed worth to keep contact with.
 #[derive(Clone, Debug)]
-pub struct Link {
+pub(crate) struct Link {
     peer: Peer,
     endpoint: Endpoint,
     connections: Arc<RwLock<BTreeMap<ConnId, Connection>>>,
@@ -46,43 +39,38 @@ impl Link {
         }
     }
 
-    #[allow(unused)]
-    pub(crate) fn name(&self) -> XorName {
-        self.peer.name()
-    }
-
-    pub async fn send_bi(
+    pub(crate) async fn send_bi(
         &self,
         bytes: UsrMsgBytes,
         msg_id: MsgId,
-    ) -> Result<RecvStream, SendToOneError> {
-        debug!("sending bidi msg out... {msg_id:?} to {:?} ", self.peer);
+    ) -> Result<RecvStream, LinkError> {
+        let peer = self.peer;
+        debug!("sending bidi msg out... {msg_id:?} to {peer:?}");
         let conn = self.get_or_connect(msg_id).await?;
-        debug!("connection got {msg_id:?} to {:?}", self.peer);
-        let (mut send_stream, recv_stream) =
-            conn.open_bi().await.map_err(SendToOneError::Connection)?;
+        debug!("connection got {msg_id:?} to {peer:?}");
+        let (mut send_stream, recv_stream) = conn.open_bi().await.map_err(LinkError::Connection)?;
 
         debug!("{msg_id:?} bidi opened");
         send_stream.set_priority(10);
         send_stream
             .send_user_msg(bytes.clone())
             .await
-            .map_err(SendToOneError::Send)?;
+            .map_err(LinkError::Send)?;
         debug!("{msg_id:?} bidi msg sent");
 
-        send_stream.finish().await.or_else(|err| match err {
+        send_stream.finish().await.or_else(|error| match error {
             qp2p::SendError::StreamLost(qp2p::StreamError::Stopped(_)) => Ok(()),
-            _ => Err(SendToOneError::Send(err)),
+            _ => Err(LinkError::Send(error)),
         })?;
         debug!("{msg_id:?} bidi finished");
         Ok(recv_stream)
     }
 
     // Get a connection or create a fresh one
-    async fn get_or_connect(&self, msg_id: MsgId) -> Result<Connection, SendToOneError> {
+    async fn get_or_connect(&self, msg_id: MsgId) -> Result<Connection, LinkError> {
         debug!("Attempting to get conn read lock... {msg_id:?}");
         let empty_conns = self.connections.read().await.is_empty();
-        debug!("lockgottt {msg_id:?}");
+        debug!("lock got {msg_id:?}");
         if empty_conns {
             // read again
             // first caller will find none again, but the subsequent callers
@@ -102,12 +90,6 @@ impl Link {
         }
     }
 
-    /// Is this Link currently connected?
-    pub(crate) async fn is_connected(&self) -> bool {
-        // self.remove_expired().await;
-        !self.connections.read().await.is_empty()
-    }
-
     /// Uses qp2p to create a connection and stores it on Self.
     /// Returns early without creating a new connection if an existing connection is found (which may have been created before we can get the write lock).
     ///
@@ -116,11 +98,12 @@ impl Link {
     pub(crate) async fn create_connection_if_none_exist(
         &self,
         msg_id: Option<MsgId>,
-    ) -> Result<Connection, SendToOneError> {
+    ) -> Result<Connection, LinkError> {
+        let peer = self.peer;
         // grab write lock to prevent many many conns being opened at once
-        debug!("[CONN WRITE]: {msg_id:?} to {:?}", self.peer);
+        debug!("[CONN WRITE]: {msg_id:?} to {peer:?}");
         let mut conns = self.connections.write().await;
-        debug!("[CONN WRITE]: lock obtained {msg_id:?} to {:?}", self.peer);
+        debug!("[CONN WRITE]: lock obtained {msg_id:?} to {peer:?}");
 
         // let's double check we havent got a connection meanwhile
         if let Some(conn) = conns.iter().next().map(|(_, c)| c.clone()) {
@@ -128,14 +111,14 @@ impl Link {
             return Ok(conn);
         }
 
-        debug!("{msg_id:?} creating connnnn to {:?}", self.peer);
+        debug!("{msg_id:?} creating connnnn to {peer:?}");
         let (conn, _incoming_msgs) = self
             .endpoint
-            .connect_to(&self.peer.addr())
+            .connect_to(&peer.addr())
             .await
-            .map_err(SendToOneError::Connection)?;
+            .map_err(LinkError::Connection)?;
 
-        debug!("{msg_id:?} conn creating done {:?}", self.peer);
+        debug!("{msg_id:?} conn creating done {peer:?}");
         trace!(
             "{} to {} (id: {})",
             LogMarker::ConnectionOpened,
@@ -149,28 +132,14 @@ impl Link {
     }
 }
 
-/// Errors that can be returned from `Comm::send_to_one`.
 #[derive(Debug)]
-pub enum SendToOneError {
-    ///
+/// Errors returned when using a Link
+pub enum LinkError {
+    /// Failed to connect to a peer
     Connection(qp2p::ConnectionError),
-    ///
+    /// Failed to send a msg to a peer
     Send(qp2p::SendError),
-    #[cfg(feature = "chaos")]
+    #[cfg(features = "chaos")]
     /// ChaosNoConn
     ChaosNoConnection,
-}
-
-impl SendToOneError {
-    ///
-    #[allow(unused)]
-    pub(crate) fn is_local_close(&self) -> bool {
-        matches!(
-            self,
-            Self::Connection(qp2p::ConnectionError::Closed(qp2p::Close::Local))
-                | Self::Send(qp2p::SendError::ConnectionLost(
-                    qp2p::ConnectionError::Closed(qp2p::Close::Local)
-                ))
-        )
-    }
 }
